refactor: 의미를 명확히 하기 위해 파일명 및 변수명을 변경한다
